+++ test/GitDataTest.php
@@ -22,7 +22,7 @@ class GitDataTest extends TestCase {
       $data = GitData::fromJson([]);
       expect($data)->to->be->instanceOf(GitData::class);
       expect($data->getBranch())->to->be->empty;
-      expect($data->getCommit())->to->be->null;
+      expect($data->getCommit())->to->be->instanceOf(GitCommit::class);
       expect($data->getRemotes())->to->be->empty;
     });
 
@@ -80,10 +80,10 @@ class GitDataTest extends TestCase {
    */
   public function testJsonSerialize() {
     it('should return a map with default values for a newly created instance', function() {
-      $map = (new GitData)->jsonSerialize();
+      $map = (new GitData(new GitCommit('')))->jsonSerialize();
       expect(get_object_vars($map))->to->have->lengthOf(3);
       expect($map->branch)->to->be->empty;
-      expect($map->head)->to->be->null;
+      expect($map->head)->to->be->instanceOf(\stdClass::class);
       expect($map->remotes)->to->be->an('array')->and->be->empty;
     });